FastLocalCorrelationCoefficients

Documentation for FastLocalCorrelationCoefficients.

FastLocalCorrelationCoefficients.flcc

FastLocalCorrelationCoefficients.flccComp

FastLocalCorrelationCoefficients.flccPrec

FastLocalCorrelationCoefficients.lcc

Computing local correlation coefficients (also known as LCCs) is a basic step in various image-based data or information processing applications, including template or pattern matching, detection and estimation of motion or some other change in an image frame series, image registration from data collected at different times, projections, perspectives or with different acquisition modalities, and compression across multiple image frames.

The Fast Local Correlation Coefficients (FLCC) Library FastLocalCorrelationCoefficients.jl computes the Correlation Coefficients with Local Normalization for arbitrary dimensional tensors with real or complex values.

For more information see:

X. Sun, N. P. Pitsianis, and P. Bientinesi, Fast computation of local correlation coefficients, Proc. SPIE 7074, 707405 (2008)

G. Papamakarios, G. Rizos, N. P. Pitsianis, and X. Sun, Fast computation of local correlation coefficients on graphics processing units, Proc. SPIE 7444, 744412 (2009)

FastLocalCorrelationCoefficients.flcc — Method

  flcc(haystack,needle)

Calculate the local correlation coefficients fast using fft.

Example

Suppose you have a haystack, a tensor of reals and a needle, a smaller tensor of the same dimensionality that you are are trying to locate in the haystack. Note that the needle might be scaled and translated.

The position of the maximum element of LCC is the best match between the needle and a sliding window of haystack

julia> using FastLocalCorrelationCoefficients

julia> haystack = rand(2^10,2^10);

julia> needle = rand(1) .* haystack[42:48, 45:50] .+ rand(1);

julia> LCC = flcc(haystack,needle);

julia> argmax(LCC)
CartesianIndex(42, 45)

FastLocalCorrelationCoefficients.flccComp — Method

Then, use the precomputed value tuple for every needle.

  haystack = rand(2^20)
  needle1 = rand(1) .* haystack[2:8] .+ rand(1)
  needle2 = rand(1) .* haystack[42:48] .+ rand(1)
  needle3 = rand(1) .* haystack[end-6:end] .+ rand(1)
  precomp = FastLocalCorrelationCoefficients.flccPrec(haystack,size(needle1))
  argmax(FastLocalCorrelationCoefficients.flccComp(precomp,needle1)) == 2
  argmax(FastLocalCorrelationCoefficients.flccComp(precomp,needle2)) == 42
  argmax(FastLocalCorrelationCoefficients.flccComp(precomp,needle3)) == 2^20-6

FastLocalCorrelationCoefficients.flccPrec — Method

When you need to search for several needles of the same size, then you can avoid redundant computations by precomputing all common information.

  haystack = rand(2^20)
  needle1 = rand(1) .* haystack[2:8] .+ rand(1)
  needle2 = rand(1) .* haystack[42:48] .+ rand(1)
  needle3 = rand(1) .* haystack[end-6:end] .+ rand(1)
  precomp = FastLocalCorrelationCoefficients.flccPrec(haystack,size(needle1))
  argmax(FastLocalCorrelationCoefficients.flccComp(precomp,needle1)) == 2
  argmax(FastLocalCorrelationCoefficients.flccComp(precomp,needle2)) == 42
  argmax(FastLocalCorrelationCoefficients.flccComp(precomp,needle3)) == 2^20-6

FastLocalCorrelationCoefficients.lcc — Method

  lcc(haystack,needle)

Calculate the local correlation coefficients directly.

Example

Suppose you have a haystack, a tensor of reals and a needle, a smaller tensor of the same dimensionality that you are are trying to locate in the haystack. Note that the needle might be scaled and translated.

The position of the maximum element of LCC is the best match between the needle and a sliding window of haystack

julia> using FastLocalCorrelationCoefficients

julia> haystack = rand(2^10,2^10);

julia> needle = rand(1) .* haystack[42:48, 45:50] .+ rand(1);

julia> LCC = lcc(haystack,needle);

julia> argmax(LCC)
CartesianIndex(42, 45)